backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pandas==2.1.4
openpyxl==3.1.2
supabase==2.3.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
ruff==0.1.14
pytest==7.4.4
httpx==0.26.0
